+++ nodes/__init__.py
@@ -1,6 +1,6 @@
 """Node classes for udi-plugin-konnected."""
 
-VERSION = "1.0.0"
+VERSION = "1.0.1"
 
 from .Controller import Controller
 from .GarageDoor import GarageDoor